requirements.txt
beautifulsoup4~=4.0
black~=22.0
boto3~=1.0
github3.py~=3.0
jedi
lxml~=4.0
matplotlib~=3.0
neovim
numpy~=1.0
pandas~=1.0
pip-tools
pudb
pydantic~=1.0
pynvim
pytest-cov~=4.0
pytest~=7.0
pipx
poetry